crossfit_box/requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
